Add human scGrapHiC fine-tuning pipeline and infrastructure
Major additions:
- Fine-tuning scripts for CD34+ hematopoietic cells with bulk Hi-C
- Multiple NPZ generation scripts (v2-v4) for different bulk Hi-C sources
- Train/validation/test split creation with chromosome-based stratification
- Comprehensive batch scripts for data processing, training, and inference
- Inference pipeline for full chromosome evaluation
- Analysis tools: sparsity checking, bulk mismatch detection, feature inspection
- Documentation: finetuning guide, augmentation workflow, inference plan

Key files:
- finetune_human.py: Main fine-tuning script with GM12878 bulk Hi-C
- create_human_npz_v2.py: NPZ generation with configurable bulk sources
- create_train_test_val_split.py: Chromosome-stratified data splitting
- batch_scripts/: Complete SLURM job submission pipeline
- compare_distributions.py: Analyze chromosome/cell-type distributions

Model performance:
- GM12878 fine-tuning: 22.63% validation SCC (chr9,12,17), 17% full genome SCC
- Dataset: 2,815 CD34+ cells (HSC, MPP, LMPP) from GSE238001
- Resolution: 50kb bins, library_size=25000

+++ src/graph_pe.py
@@ -103,8 +103,9 @@ def graph_pe(matrix, encoding_dim, lap_norm='sym', eig_norm='abs-max'):
 
     
     
-    sparse_matrix = sparse.csr_matrix(matrix)
+    sparse_matrix = sparse.csr_matrix(matrix.astype(np.float32))
     edge_index, edge_weight = from_scipy_sparse_matrix(sparse_matrix)
+    edge_weight = edge_weight.float()  # Ensure float type for laplacian
     
     edge_index, edge_weight = get_laplacian(edge_index, edge_weight, normalization=lap_norm, num_nodes=N)
     L = to_scipy_sparse_matrix(edge_index, edge_weight, N)
